+++ include/render/Integrator.h
@@ -78,6 +78,8 @@ namespace render {
     std::string intersectionBackendAvailability;
     std::string intersectionBackendFallbackReason;
     std::string intersectionBackendExecutionPath;
+    std::string intersectionBackendClosestHitExecutionPath;
+    std::string intersectionBackendAnyHitExecutionPath;
     bool intersectionBackendPlatformGpuDeviceAvailable{false};
     bool intersectionBackendPlatformGpuRenderPathAvailable{false};
     bool intersectionSceneCompiled{false};